Move SplitController out of components folder
